Get rid of the Attributes concept and its implementations.
The Attributes was conceptually useful for POSIX systems.  The class itself
can configure the execution environment to launch a new process because it
simply needs to do the required changes in between the fork and exec calls.

Unfortunately, this is not extensible to Win32 systems because child
processes are spawned in a completely different way.  Providing a
win32_attributes class could be useless without also providing a matching
win32_launcher that knew how to manage the extra attributes.

So simply merge the attributes functionality into the launcher.  This also
simplifies things a lot, making them less confusing.

There are more details in the DesignThoughts wiki page.