Get rid of the launcher flags to set up redirections and instead use multiple parameters, one per data flow.
This is to improve flexibility.  The previous interface was too limited: the
user could only redirect a stream or leave it "undefined" -- where undefined
means that it was not specified whether it'd remain open or inherit its
parent handle.

I want the user to be able to exactly specify how each stream shall behave:
closed, silent, inherited or redirected (and maybe I'm missing some other
state).  At the moment, only closed and redirected are provided (to keep the
API change mechanical and isolated from functionality extensions), although
only the latter is implemented correctly.

The only problem I see in this change is that creating a launcher object now
requires too much typing.  I can think of some ways to simplify this (macros
or convenience constructors for typical calls), but this will be done later
on once the full-featured interface is well defined.