Move the command line shell execution test from launcher_test to command_line_test.cpp.  It makes more sense to keep it here.
While doing this, introduce a variable that indicates the current API and
use it in command_line_test.cpp (for now).  This is to avoid conditional
compilation where it is really not necessary, forcing all the code to
be compiled in more platforms and compilers; some build errors might be
caught this way.