Start of the huge refactoring: - Drop the 'launcher' (alongside 'launcher_base') class.  These are now   replaced by a 'context' class that holds only information about the   execution context and a 'launch' free function thus decoupling the two   concepts. - Make the 'environment' class public now that it is exposed through   'context'. - Simplify 'environment' a lot by using a plain standard container.   This is "less magic" and more on par with the new command line management. - Reorganize code, tests and examples a lot to suit all the changes above. - Possibly something else I forget.
There is still a lot of broken stuff: none of the OS-specific classes work
now.  These will come later, but the changes are already big enough to go in.

Documentation is also left untouched as it will require a serious rewrite.
Shouldn't be done until the new API is stabilized.